chore(DTFS2-6934): refactor bss deals/facilities and address code review comments
