+++ utils/mock-data-loader/bss/deals/deals-to-not-submit-to-tfm/2.js
@@ -1,11 +1,8 @@
-const { nowTimestamp } = require('../dates');
-const MANDATORY_CRITERIA = require('../mandatoryCriteria');
-const { MAKER } = require('../../portal/roles');
+const { nowTimestamp } = require('../../dates');
+const { MAKER } = require('../../../portal/roles');
+const { getLatestMandatoryCriteria } = require('../../helpers/getLatestMandatoryCriteria');
 
-// de-structure and create a new array, so `sort` doesn't impact 'MANDATORY_CRITERIA'
-let mandatoryCriteria = [...MANDATORY_CRITERIA].sort((a, b) => (a.version > b.version ? 1 : -1));
-// get the latest mandatory criteria (sorted by version)
-mandatoryCriteria = mandatoryCriteria[mandatoryCriteria.length - 1];
+const mandatoryCriteria = getLatestMandatoryCriteria();
 
 module.exports = {
   mockId: 2,
