+++ utils/mock-data-loader/bss/deals/deals-to-not-submit-to-tfm/6.js
@@ -1,19 +1,16 @@
-const { nowTimestamp } = require('../dates');
-const MANDATORY_CRITERIA = require('../mandatoryCriteria');
-const { MAKER } = require('../../portal/roles');
+const { nowTimestamp } = require('../../dates');
+const { MAKER } = require('../../../portal/roles');
+const { getLatestMandatoryCriteria } = require('../../helpers/getLatestMandatoryCriteria');
 
-// de-structure and create a new array, so `sort` doesn't impact 'MANDATORY_CRITERIA'
-let mandatoryCriteria = [...MANDATORY_CRITERIA].sort((a, b) => (a.version > b.version ? 1 : -1));
-// get the latest mandatory criteria (sorted by version)
-mandatoryCriteria = mandatoryCriteria[mandatoryCriteria.length - 1];
+const mandatoryCriteria = getLatestMandatoryCriteria();
 
 module.exports = {
-  mockId: 9,
+  mockId: 6,
   dealType: 'BSS/EWCS',
   submissionType: 'Manual Inclusion Application',
   updatedAt: Date.now(),
-  bankInternalRefName: 'Manual Test 7',
-  additionalRefName: 'Manual Test 7',
+  bankInternalRefName: 'Manual Test 4',
+  additionalRefName: 'Manual Test 4',
   bank: {
     id: '9',
     name: 'UKEF test bank (Delegated)',
@@ -23,7 +20,6 @@ module.exports = {
   },
   details: {
     status: 'Draft',
-    submissionType: 'Manual Inclusion Application',
     maker: {
       _id: '60f7d72654f99900074c0a6d',
       username: 'maker1@ukexportfinance.gov.uk',
@@ -205,7 +201,7 @@ module.exports = {
   comments: [],
   editedBy: [
     {
-      date: '1626972928399',
+      date: '1626972461251',
       username: 'maker1@ukexportfinance.gov.uk',
       roles: [
         MAKER
