+++ utils/mock-data-loader/bss/deals/deals-to-not-submit-to-tfm/index.js
@@ -1,0 +1,14 @@
+/* eslint-disable global-require */
+module.exports = [
+  require('./1'),
+  require('./2'),
+  require('./3'),
+  require('./4'),
+  require('./5'),
+  require('./6'),
+  require('./7'),
+  require('./8'),
+  require('./9'),
+  require('./10'),
+  require('./11'),
+];
