+++ utils/mock-data-loader/bss/facilities/facilities-to-not-submit-to-tfm/deal-1-facilities.js
@@ -1,14 +1,14 @@
 const {
   nowTimestamp, twoMonths, twoMonthsTimestamp, threeMonths,
-} = require('../dates');
+} = require('../../dates');
 
 module.exports = [
   {
-    mockDealId: 3,
+    mockDealId: 1,
     type: 'Bond',
     createdDate: nowTimestamp,
     bondIssuer: '',
-    bondType: 'Maintenance bond',
+    bondType: 'Performance bond',
     ukefGuaranteeInMonths: null,
     facilityStage: 'Issued',
     hasBeenIssued: true,
@@ -18,11 +18,11 @@ module.exports = [
     'coverEndDate-day': threeMonths.day,
     'coverEndDate-month': threeMonths.month,
     'coverEndDate-year': threeMonths.year,
-    name: 'Test Bond 3',
+    name: 'Test Bond 1',
     bondBeneficiary: '',
     requestedCoverStartDate: twoMonthsTimestamp,
     updatedAt: nowTimestamp,
-    value: '600000.00',
+    value: '500000.00',
     currencySameAsSupplyContractCurrency: 'true',
     currency: {
       text: 'GBP - UK Sterling',
@@ -34,12 +34,12 @@ module.exports = [
     'conversionRateDate-month': null,
     'conversionRateDate-year': null,
     riskMarginFee: '2',
-    coveredPercentage: '60',
+    coveredPercentage: '80',
     minimumRiskMarginFee: '',
     guaranteeFeePayableByBank: '1.8000',
-    ukefExposure: '360,000.00',
-    feeFrequency: 'Quarterly',
+    ukefExposure: '400,000.00',
+    feeFrequency: 'Monthly',
     feeType: 'In arrear',
-    dayCountBasis: '360',
+    dayCountBasis: '365',
   },
 ];
