+++ utils/mock-data-loader/bss/facilities/facilities-to-not-submit-to-tfm/deal-10-facilities.js
@@ -1,6 +1,6 @@
 const {
   nowTimestamp, twoMonths, twoMonthsTimestamp, threeMonths,
-} = require('../dates');
+} = require('../../dates');
 
 module.exports = [
   {
