+++ utils/mock-data-loader/bss/facilities/facilities-to-not-submit-to-tfm/deal-5-facilities.js
@@ -1,55 +1,16 @@
 const {
   nowTimestamp, twoMonths, twoMonthsTimestamp, threeMonths,
-} = require('../dates');
+} = require('../../dates');
 
 module.exports = [
   {
-    mockDealId: 6,
-    type: 'Bond',
-    ukefGuaranteeInMonths: null,
-    facilityStage: 'Issued',
-    hasBeenIssued: true,
-    'coverEndDate-day': threeMonths.day,
-    'coverEndDate-month': threeMonths.month,
-    'coverEndDate-year': threeMonths.year,
-    name: 'Test Bond 6',
-    requestedCoverStartDate: twoMonthsTimestamp,
-    value: '600000.00',
-    currencySameAsSupplyContractCurrency: 'true',
-    currency: {
-      text: 'GBP - UK Sterling',
-      id: 'GBP',
-      currencyId: 12,
-    },
-    conversionRate: null,
-    'conversionRateDate-day': null,
-    'conversionRateDate-month': null,
-    'conversionRateDate-year': null,
-    createdDate: nowTimestamp,
-    bondIssuer: '',
-    bondType: 'Maintenance bond',
-    'requestedCoverStartDate-day': twoMonths.day,
-    'requestedCoverStartDate-month': twoMonths.month,
-    'requestedCoverStartDate-year': twoMonths.year,
-    bondBeneficiary: '',
-    ukefExposure: '360,000.00',
-    updatedAt: nowTimestamp,
-    riskMarginFee: '2',
-    coveredPercentage: '60',
-    minimumRiskMarginFee: '',
-    guaranteeFeePayableByBank: '1.8000',
-    feeFrequency: 'Quarterly',
-    feeType: 'In arrear',
-    dayCountBasis: '360',
-  },
-  {
-    mockDealId: 6,
+    mockDealId: 5,
     type: 'Loan',
     ukefGuaranteeInMonths: '24',
     'coverEndDate-day': null,
     'coverEndDate-month': null,
     'coverEndDate-year': null,
-    name: 'Test Loan 6',
+    name: 'Test Loan 5',
     requestedCoverStartDate: null,
     disbursementAmount: null,
     value: '100000.00',
@@ -78,5 +39,44 @@ module.exports = [
     premiumFrequency: 'Monthly',
     premiumType: 'In advance',
     dayCountBasis: '365',
+  },
+  {
+    mockDealId: 5,
+    type: 'Bond',
+    ukefGuaranteeInMonths: null,
+    facilityStage: 'Issued',
+    hasBeenIssued: true,
+    'coverEndDate-day': threeMonths.day,
+    'coverEndDate-month': threeMonths.month,
+    'coverEndDate-year': threeMonths.year,
+    name: 'Test Bond 5',
+    requestedCoverStartDate: twoMonthsTimestamp,
+    value: '600000.00',
+    currencySameAsSupplyContractCurrency: 'true',
+    currency: {
+      text: 'GBP - UK Sterling',
+      id: 'GBP',
+      currencyId: 12,
+    },
+    conversionRate: null,
+    'conversionRateDate-day': null,
+    'conversionRateDate-month': null,
+    'conversionRateDate-year': null,
+    createdDate: nowTimestamp,
+    bondIssuer: '',
+    bondType: 'Maintenance bond',
+    'requestedCoverStartDate-day': twoMonths.day,
+    'requestedCoverStartDate-month': twoMonths.month,
+    'requestedCoverStartDate-year': twoMonths.year,
+    bondBeneficiary: '',
+    ukefExposure: '360,000.00',
+    updatedAt: nowTimestamp,
+    riskMarginFee: '2',
+    coveredPercentage: '60',
+    minimumRiskMarginFee: '',
+    guaranteeFeePayableByBank: '1.8000',
+    feeFrequency: 'Quarterly',
+    feeType: 'In arrear',
+    dayCountBasis: '360',
   },
 ];
