+++ utils/mock-data-loader/bss/facilities/facilities-to-not-submit-to-tfm/deal-6-facilities.js
@@ -1,10 +1,10 @@
 const {
   nowTimestamp, twoMonths, twoMonthsTimestamp, threeMonths,
-} = require('../dates');
+} = require('../../dates');
 
 module.exports = [
   {
-    mockDealId: 2,
+    mockDealId: 6,
     type: 'Bond',
     ukefGuaranteeInMonths: null,
     facilityStage: 'Issued',
@@ -12,9 +12,9 @@ module.exports = [
     'coverEndDate-day': threeMonths.day,
     'coverEndDate-month': threeMonths.month,
     'coverEndDate-year': threeMonths.year,
-    name: 'Test Bond 2',
+    name: 'Test Bond 6',
     requestedCoverStartDate: twoMonthsTimestamp,
-    value: '500000.00',
+    value: '600000.00',
     currencySameAsSupplyContractCurrency: 'true',
     currency: {
       text: 'GBP - UK Sterling',
@@ -27,33 +27,32 @@ module.exports = [
     'conversionRateDate-year': null,
     createdDate: nowTimestamp,
     bondIssuer: '',
-    bondType: 'Performance bond',
+    bondType: 'Maintenance bond',
     'requestedCoverStartDate-day': twoMonths.day,
     'requestedCoverStartDate-month': twoMonths.month,
     'requestedCoverStartDate-year': twoMonths.year,
     bondBeneficiary: '',
-    ukefExposure: '400,000.00',
+    ukefExposure: '360,000.00',
     updatedAt: nowTimestamp,
     riskMarginFee: '2',
-    coveredPercentage: '80',
+    coveredPercentage: '60',
     minimumRiskMarginFee: '',
     guaranteeFeePayableByBank: '1.8000',
-    feeFrequency: 'Monthly',
+    feeFrequency: 'Quarterly',
     feeType: 'In arrear',
-    dayCountBasis: '365',
-    viewedPreviewPage: true,
+    dayCountBasis: '360',
   },
   {
-    mockDealId: 2,
+    mockDealId: 6,
     type: 'Loan',
-    ukefGuaranteeInMonths: '12',
+    ukefGuaranteeInMonths: '24',
     'coverEndDate-day': null,
     'coverEndDate-month': null,
     'coverEndDate-year': null,
-    name: 'Test Loan 2',
+    name: 'Test Loan 6',
     requestedCoverStartDate: null,
     disbursementAmount: null,
-    value: '250000.00',
+    value: '100000.00',
     currencySameAsSupplyContractCurrency: 'true',
     currency: {
       text: 'GBP - UK Sterling',
@@ -70,10 +69,10 @@ module.exports = [
     'requestedCoverStartDate-day': null,
     'requestedCoverStartDate-month': null,
     'requestedCoverStartDate-year': null,
-    ukefExposure: '200,000.00',
+    ukefExposure: '60,000.00',
     updatedAt: nowTimestamp,
     interestMarginFee: '2',
-    coveredPercentage: '80',
+    coveredPercentage: '60',
     minimumQuarterlyFee: '',
     guaranteeFeePayableByBank: '1.8000',
     premiumFrequency: 'Monthly',
