+++ utils/mock-data-loader/bss/facilities/facilities-to-not-submit-to-tfm/deal-7-facilities.js
@@ -1,14 +1,16 @@
-const { nowTimestamp, twoMonths, twoMonthsTimestamp } = require('../dates');
+const {
+  nowTimestamp, twoMonths, twoMonthsTimestamp, threeMonths,
+} = require('../../dates');
 
 module.exports = [
   {
-    mockDealId: 4,
+    mockDealId: 7,
     type: 'Loan',
     ukefGuaranteeInMonths: '24',
     'coverEndDate-day': null,
     'coverEndDate-month': null,
     'coverEndDate-year': null,
-    name: 'Test Loan 4',
+    name: 'Test Loan 7',
     requestedCoverStartDate: null,
     disbursementAmount: null,
     value: '100000.00',
@@ -39,15 +41,18 @@ module.exports = [
     dayCountBasis: '365',
   },
   {
-    mockDealId: 4,
+    mockDealId: 7,
     type: 'Bond',
     ukefGuaranteeInMonths: null,
     facilityStage: 'Issued',
     hasBeenIssued: true,
-    'coverEndDate-day': '22',
-    'coverEndDate-month': '09',
-    'coverEndDate-year': '2023',
-    name: 'Test Bond 4',
+    'requestedCoverStartDate-day': twoMonths.day,
+    'requestedCoverStartDate-month': twoMonths.month,
+    'requestedCoverStartDate-year': twoMonths.year,
+    'coverEndDate-day': threeMonths.day,
+    'coverEndDate-month': threeMonths.month,
+    'coverEndDate-year': threeMonths.year,
+    name: 'Test Bond 7',
     requestedCoverStartDate: twoMonthsTimestamp,
     value: '600000.00',
     currencySameAsSupplyContractCurrency: 'true',
@@ -63,9 +68,6 @@ module.exports = [
     createdDate: nowTimestamp,
     bondIssuer: '',
     bondType: 'Maintenance bond',
-    'requestedCoverStartDate-day': twoMonths.day,
-    'requestedCoverStartDate-month': twoMonths.month,
-    'requestedCoverStartDate-year': twoMonths.year,
     bondBeneficiary: '',
     ukefExposure: '360,000.00',
     updatedAt: nowTimestamp,
