+++ utils/mock-data-loader/bss/facilities/facilities-to-not-submit-to-tfm/deal-8-facilities.js
@@ -1,4 +1,4 @@
-const { nowTimestamp } = require('../dates');
+const { nowTimestamp } = require('../../dates');
 
 module.exports = [
   {
