+++ utils/mock-data-loader/bss/facilities/index.js
@@ -1,15 +1,5 @@
 /* eslint-disable global-require */
 module.exports = [
-  ...require('./deal-1-facilities'),
-  ...require('./deal-2-facilities'),
-  ...require('./deal-3-facilities'),
-  ...require('./deal-4-facilities'),
-  ...require('./deal-5-facilities'),
-  ...require('./deal-6-facilities'),
-  ...require('./deal-7-facilities'),
-  ...require('./deal-8-facilities'),
-  ...require('./deal-9-facilities'),
-  ...require('./deal-10-facilities'),
-  ...require('./deal-11-facilities'),
+  ...require('./facilities-to-not-submit-to-tfm'),
   ...require('./facilities-to-submit-to-tfm'),
 ];
