--- utils/mock-data-loader/bss/toJson.js
@@ -1,5 +1,0 @@
-const mocks = require('./index');
-
-const deal = JSON.stringify(mocks.DEALS.dealsNotToSubmit[0], null, 2);
-
-console.info(deal);
