+++ utils/mock-data-loader/package.json
@@ -13,6 +13,7 @@
     "generate-password": "^1.7.1",
     "json2csv": "^5.0.7",
     "jsonwebtoken": "^9.0.2",
+    "lodash": "^4.17.21",
     "moment": "^2.29.4",
     "mongodb": "^6.3.0"
   },
